User can now create a message

+++ mobile/src/main/java/sketchagram/chalmers/com/model/AMessage.java
@@ -6,11 +6,11 @@ import java.util.List;
  * Created by Bosch on 10/02/15.
  */
 public abstract class AMessage {
-    private final double TIMESTAMP;
+    private final long TIMESTAMP;
     private final ADigitalPerson SENDER;
     private final List<ADigitalPerson> RECEIVER;
 
-    protected AMessage(double timestamp, ADigitalPerson sender, List<ADigitalPerson> receiver) {
+    protected AMessage(long timestamp, ADigitalPerson sender, List<ADigitalPerson> receiver) {
         this.TIMESTAMP = timestamp;
         this.SENDER = sender;
         this.RECEIVER = receiver;
